[W2-T09B] Add TransferService with ownership, budget, and lock checks
- Add TransferService + TransferServiceImpl (service package)
- Add minimal placeholder PlayerDAO and FantasyTeamDAO (dao package),
  scoped only to what transfer logic needs; flagged as placeholders
  pending dedicated tickets for full implementations
- Add minimal placeholder SquadValidationService (service package)
  for player-availability and budget checks
- Lock-window check currently uses Transfer.transferWindowStatus /
  FantasyTeam.isLocked since fixture-based locking isn't wired up yet
- All SQL stays in DAO classes; service throws ApplicationException
  subclasses (AuthorisationException, BusinessRuleException,
  ResourceNotFoundException, ValidationException) for clear error
  handling at the Resource layer

+++ src/main/java/com/vzap/trytons/dto/TransferResponse.java
@@ -3,6 +3,7 @@ package com.vzap.trytons.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -16,6 +17,6 @@ public class TransferResponse {
     private final LocalDateTime transferDate;
     private final boolean penaltyApplied;
     private final int penaltyPointAmount;
-    private final double newRemainingBudget;
-    private final double newTotalTeamValue;
+    private final BigDecimal newRemainingBudget;
+    private final BigDecimal newTotalTeamValue;
 }
